Implement SQLite database with Dapper and fix decimal handling

+++ Expense.cs
@@ -4,7 +4,7 @@ public class Expense
 {
     public int Id { get; set; }
     public string Description { get; set; } = string.Empty;
-    public decimal Amount { get; set; }
+    public double Amount { get; set; } // Double
     public string Category { get; set; } = string.Empty;
     public DateTime Date { get; set; }
 }
